block entity support for 1.9-1.16
